version 0.1.0
Complete change of the files implementation method : you can now define the name of the files in the navbar, define which language you want among dozens and define the content of each file. Moreover, you are no longer limited by a maximum number of files, you can put as many as you want!